+++ src/keymap.rs
@@ -16,7 +16,7 @@
 #[derive(Copy, Eq, Hash, PartialEq)]
 #[cfg_attr(test, derive(Debug))]
 pub enum Key {
-  Fn{num: int, mods: KeyMod},
+  Fn{num: isize, mods: KeyMod},
   Sym{sym: KeySym, mods: KeyMod},
   Unicode{codepoint: char, mods: KeyMod},
 }
